Extract jwk validation to its own bean
