+++ security-jwt/src/main/java/io/micronaut/security/token/jwt/signature/jwks/JwkValidator.java
@@ -1,0 +1,36 @@
+/*
+ * Copyright 2017-2019 original authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package io.micronaut.security.token.jwt.signature.jwks;
+
+import com.nimbusds.jose.jwk.JWK;
+import com.nimbusds.jwt.SignedJWT;
+
+/**
+ * Validates a JWT signature with a JSON Web Key (JWK).
+ *
+ * @author Sergio del Amo
+ * @since 1.1.0
+ */
+public interface JwkValidator {
+
+    /**
+     *
+     * @param jwt Signed JWT
+     * @param jwk JSON Web Key
+     * @return true if signature is verified
+     */
+    boolean validate(SignedJWT jwt, JWK jwk);
+}
